skills/gemini-image/requirements.txt
# Gemini Image Generator 의존 패키지
google-genai>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
